feat: Create Excel report
cleanup report

Issue #13
