+++ pmlib/types.py
@@ -15,6 +15,8 @@
 #
 #    Copyright (C) 2017, Kai Raphahn <kai.raphahn@laburec.de>
 #
+import abc
+from abc import ABCMeta
 
 from dataclasses import dataclass, field
 from enum import Enum
@@ -32,8 +34,12 @@ __all__ = [
     "Navigation",
     "Counter",
     "EntryData",
-    "Position"
+    "Position",
+    "Report"
 ]
+
+import pmlib
+from pmlib.item import Item
 
 
 class Entry(Enum):
@@ -202,3 +208,14 @@ class Position(object):
     @property
     def length(self) -> int:
         return self.end - self.start
+
+
+class Report(metaclass=ABCMeta):
+
+    def __init__(self):
+        self.root: Item = pmlib.data.root
+        return
+
+    @abc.abstractmethod
+    def create(self) -> bool:
+        pass
